Rename files in `csrc/serde` directory. (#1496)
This PR renames files in the `csrc/serde` directory.
