+++ csrc/serde/fusion_record.h
